// api/src/especialidade-medica/dto/resposta-especialidade-medica.dto.ts
/** Resposta detalhe — BigInt como string no JSON. */
export class RespostaEspecialidadeMedicaDto {
  id: string;
  id_cbo: string | null;
  descricao: string | null;
  codigo_ipsemg: string | null;
  codigo_externo: string | null;
  codigo_migracao: string | null;
  id_usuario_auditoria: string | null;
  endereco_ip_auditoria: string | null;
  nome_aplicacao_auditoria: string | null;
}